Internals

Internals

PGFPlotsX.print_tex — Function.

print_tex(io, elt, [container])

Print elt to io as LaTeX code. The optional third argument allows methods to work differently depending on the container.

This method should indent as if at the top level, containers indent their contents as necessary. See print_indent.

source

print_tex(io, str)

Print a string as is, terminated with a newline.

Note

This is used as a workaround for LaTeX code that does not have a corresponding type, eg as elements in Axis. raw or LaTeXStrings are useful to avoid piling up backslashes. The newline is added to separate tokens.

source

print_tex(io, x)

Real numbers are printed as is, except for non-finite representation.

source

PGFPlotsX.print_indent — Function.

print_indent(f, io_main)

Call the f with an IO buffer, capture the output, print it to io_main indended with four spaces.

source

print_indent(io_main, elt)

Print elt to io with indentation. Shortcut for the function wrapper of print_indent for a single element.

source

PGFPlotsX.expand_scanlines — Function.

expand_scanlines(n, nrow)

Expand scanlines, which is a vector of scanline positions or an integer for repeated scanlines, into a Vector{Int}.

source

PGFPlotsX.Options — Type.

Options(pairs; print_empty)

Options passed to PGFPlots for various structures (table, plot, etc).

Contents emitted in key = value form, or key when value ≡ nothing. Also see the @pgf convenience macro.

When print_empty = false (the default), empty options are not printed. Use print_empty = true to force printing a [] in this case.

source

PGFPlotsX.CUSTOM_PREAMBLE_PATH — Constant.

A file which is spliced directly to the preamble. Customize the file at this path for site-specific setting that apply for every plot.

source

PGFPlotsX.DEFAULT_PREAMBLE — Constant.

The default preamble for LaTeX documents. Don't change this, customize CUSTOM_PREAMBLE instead.

source